PFAIR: set release time of new tasks
Without a proper release time, the job will be considered "lagging
behind" for quite a while, which breaks the period enforcement. This
bug manifested only in the absence of a synchronous release (which set
a proper release time).

This patch simply sets the beginning of the next quantum as the
release time of the first job of a newly added task.