GSN-FPCA Bug Fix: v2 Race Condition at lock/unlock CP
Previous commit point "GSN-FPCA Bug Fix: Race Condition at lock/unlock
CP" describes how the bug looks like and why it happens. But previous
commit fails to fix all codes that cause the bug.

When a RT task preempt another RT task on the same CPU via cache, the
cache state of the preempted RT task is not set up until
gsnfpca_schedule() is invoked on the CPU later. However, we set the
cache state of the preempting RT task before gsnfpca_schedule() is
invoked; so we change the cache state in incorrect order.
Correct order should be: unlock the cache and then lock the cache
Our incorrect order was: lock the cache (and set the rt.cp_mask) and
unlock the cache (and clear the rt.cp_mask)

Known Issue:
We may set cache state on NO_CPU
rtspin is still in the system after the experiment.

Signed-off-by: Meng Xu