GSN-FPCA BUG FIX: Unlock cache on NO_CPU
Bug Symptom
TRACE shows that we try to unlock cache on NO_CPU

Bug Reason
We use task->linked_on to indicate which CPU the task was on and we
should unlock cache for that CPU
So we should always unlock cache before we unlink the task

Fix
The bug was caused in task_exit function.
We unlinked task before unlock the cache

Test
Run 20 RT tasks and TRACE has no such bug symptom.

Signed-off-by: Meng Xu