GSN-FPCA Check Sched Correctness
Whenever a scheduler on CPU i makes a decision to switch tasks on the
CPU, it should maintain the invariant:
The top ready task in ready queue cannot preempt the task_to_schedule on
the CPU i. The task_to_schedule is the next task decided by the
scheduler on the CPU i that will runs on the CPU i.

If the invariant is violated, we write "[ERROR]" into TRACE_TASK.

Test
Current GSN-FPCA scheduler does not have any ERROR in the trace.

Note
Scheduler does not allow priority inversion right now.

Signed-off-by: Meng Xu