GSN-FPCA Sched: Correct a transient error
Scheduler may violate the scheduling invariant for very short time.
Scheduling invariant asks that whenever after schedule() is called on a
CPU, the task on the CPU should not be preempted by the top ready_task
in the ready_queue.

We correct this transient error by using cpu_entry->linked instead of
cpu_entry->scheduled.
CPU may be in the state that a preempting task has been linked to it but
has not been scheduled on it; If a higher priority task arrives, it
should replace the preempting task just linked on the CPU and requeue
it.

Test
Passed all 13 test cases

Signed-off-by: Meng Xu