Move cache flush to context switch from sched
Cache flush can take > 1ms and should depend on the size of cache a
scheduled task will use; So it should be in the context switch logic
instead of the scheduler logic. It should be counted as context switch
overhead.

Pitfall:
    PL310 cache controller does not allow two cores to flush different
areas of cache at the same time by setting the control register. It will
cause kernel panic (Imprecise external abort)
    We use the global lock to serialize the operation, which solve the
kernel panic

Signed-off-by: Meng Xu