Fix lock issue
Issue:
We access the CP2Task map in the check_for_preemption when we do not
have the global cache controller lock.
However, we should always grab the cache controller lock before we
read/write CP2Task map.

Effect:
This issue should not crash the system. The prev cache partitions used by
the task may be incorrectly read in the check_for_preemption. since it
is a soft constaint in deciding which CP to use for a task, incorrect value
is fine.

Change:
Do not read CP2Task in check_for_preemption.
Use job_params->prev_cp instead to indicate which CPs were used by the
task.

Test:
No system crash.
Did not test the performance or evaluate if the WCET is stable after
this modification. But from the system implementaiton, we should still
have the LLC isolation.

Signed-off-by: Meng Xu