Separte the cache controller code to ARM specific code blocks. And put dummy area for supporting x86 or x86_64 based cache controller
For rt_cache.c and sched_gsn_*.c, direct calling of
"l2x0_flush_cache_ways()" to flush cache which only can be applied
to ARM PL310 cache controller is changed to general function call
"flush_cache_ways()." in cache_proc.c.

For kernel/cpu.c, there are duplicated functions if the kernel
configuration is for X86_64. So those functions are blocked.

NOTE: kernel configuration generated by "make x86_64_defconfig" is not
working at all with this system. It is required to add more
configuration from default configuration of linux distribution. And it
must make init ramdisk as well. Here is the process:

$ make x86_64_defconfig
$ cat /boot/config-3.8.0-44 >> .config
$ make menuconfig #just save then exit
$ make -j 8 bzImage modules
$ sudo make modules_install
$ sudo cp arch/x86/boot/bzImage /boot/vmlinuz-3.10.31+
$ sudo cp .config /boot/config-3.10.31+
$ sudo mkinitramfs -o /boot/initrd.img-3.10.31+ 3.10.31+
$ sudo update-grub