Transparent kernel level page coloring support
If you specify coloring value using environmental
variable (PAGE_COLORS=<color map>), memory management kernel
system allocates memory pages according to the color settings.
This mechanism only can apply on two memory allocation system calls:
brk systemcall (malloc in C and new operator in C++) and mmap
systemcall without file descriptor (i.e. fd=-1).

Let me explain briefly how to allocate memory in a process. When
a process request a memory using either brk or no-fd-mmap, kernel
tries to find free virtual address range for the requested amount
of memory. This virtual address range can be an adjcent to existing
virtual address range or totally separated range. It is noted that
brk or mmap don't allocate pages. they just mark in the task_struct
(exactly, current->mm->mmap) that some virutal address range is
used by process. The actual allocation of pages occurs when the
process tries to access(read or write) the virtual address range.
Because the kernel saved the virtual address range and no pages for
the physical memory is not assigned, the kernel causes a page fault
for the access. The page fault handler in the kernel is invoked
and it allocates pages and maps between virtual address and
physical address.

The page fault handler is handle_mm_fault() function and is located
in mm/memory.c This is the exact position to allocate color-aware
pages. If task_struct contains correct page_colors settings, this
handler tries to allocated colored pages instead of normal general
pages. Although color-aware page allocation can be applied to all
processes, we did selectively only for the valid color settings
because color-aware page allocation maintains pools for pre-allocation
and it has overhead for refilling the page pools. If we don't
specify the page coloring setting, default setting should be the
access of all color index of the cache. So there is no difference
in just bypassing to original routines.

We are using environmental variable to specify page color setting
for a process. The reason we use environmental variable is because
the color setting must be given before the process is executed.
The program loader in the kernel is the best part of initialize
the page coloring setting.

If we see the basic procedure of executing
a program in linux, a shell forks itself and call execve system call
in child process. Inside the kernel, when a process is forked,
the child process copies the parent's task_struct. Then, it
replaces the mm_struct in task_struct in execve system call.
In execve routine, basic execution information is initialized and
the kernel reads the beginning part of executing file.
The kernel tries to find the matched executor for the
executable file based on the loaded parts and passes initialized
data and file descriptor.

We modify the execve system call routine to setup coloring
information of task_struct from the environmental variable.