Active color setting using kernel boot option.
We introduce new kernel boot option: 'active_colors'
The value of the option can be hexa-decimal notation
starting with '0x'. The other possible format is a list
of activated color indexes using comma(,) and dash(-).
For example, you can set either '0xf', '0,1,2,3', or
'0-3' when you set first four color indexes for active
colors.

Once you limit the active color setting, internal page
pools which is activated will be allocated to use and
all other bits representing deactivated color indexes
from the command-line setting environment variable
"PAGE_COLORS" will be ignored. If there is no activated
color index for a process even though PAGE_COLORS setting
is specified in a process, the execution of the process
will be aborted. For example, suppose activated colors
are 0xffff (lower 16 color indexes). If you are trying
to execute a process with color setting "PAGE_COLORS=
0xffff0000" (upper 16 color indexes), it will fail to
execute because there is no active color index.

Although the boot option processing code is tested,
there can be an error for executing realtime process.

Basic procedure of run a process is like this:
1) setup expected maximum pages for the execution
 $ echo {page_count} > /proc/sys/litmus/pages_per_color
2) run a process with color setting
 $ PaGE_COLORS={color index} {application} {arguments}
3) refill the page pool for next use
 $ echo 0 > /proc/sys/litmus/refill_page_pool

a) If there is an error on procedure 1) : Lack of free
system memory. Increase the virtual machine's memory.

b) If there is an error on procedure 2) : allocated
page pool is smaller than the required amount of page
to execute the process. Try procedure 1) with increased
number of pages.

 Author:    Hyonyoung Choi <commani@gmail.com>

 On branch meng/ecrts2016/two-level-ca-sched-v2
 Changes to be committed:

	modified:   fs/exec.c
	modified:   litmus/cache_proc.c