Set default pages_per_color value as zero
When the system boots, color page pools are trying to allocate pages.
If there is no enough memory size or wrong active color mask, it causes
the kernel crash on the failure of page pool allocation. To prevent this
the default value of pages_per_color variable is set to zero.

 Author:    Hyonyoung Choi <commani@gmail.com>

	modified:   litmus/cache_proc.c