1. Add outgoing parameters of clflush start&end time in flush_cache system call. Please check the code at sys_flush_cache() in litmus.c
2. Add local invocation of clflush to test (Not used now)

3. Correct input parameter for clflush_cache_range function from page
index to virtual memory address. The value of vm_start indicates the
virtual address which is used in user space. The return value of
follow_page() is the page index of the virtual address. I passed wrong
value to clflush. (Please check the code at flush_cache_for_task() in
cache_proc.c