CLFLUSH: Add comment
Add comment on the clflush logic

Q: why the CLFLUSH latency of flushing a whole array keeps increasing as
the size of array increase?
A: Probably HW (bus, cache controller, MMU) still needs to check if the
address needs to be flush from cache or not even though the address may
not be in the cache. (Just guess)

Signed-off-by: Meng Xu