gFPca in VM: Add more debug message
The gFPca scheduler, i.e., sched_gsn_fpca2, will crash the VM when we
run the schedulder inside domU (but probably not dom0).
The crash message is:
mm.c:779:d32v0 Non-privileged (32) attempt to map I/O space 00000000

So I guess the gFPca scheduler may work on native on x86 and inside dom0;
TODO: To validate my speculation in above line

If it is because domU cannot map MMIO address, then we should figure out
which part of gFPca scheduler code does the MMIO; and then we need to
replace it for x86.
TODO: Discuss this with HY since he did the porting from ARM to x86 as
well.

Signed-off-by: Meng Xu