gFPca: working version
We set cpu_entry's cache partition when a task status changes to
CACHE_WILL_USE or CACHE_WILL_CLEAR
Whenever we switch CPUs' linked pointers to tasks, like inside the
function link_task_to_cpu(), we also need to update the cpu_entry's
cache partition setting.

Test
    Test on Intel CAT machine on bare metal
    Run 200 randomly generated tasks at the same time, No BUG is tracked
in the log, except that the get_prev_cps() may report a BUG information
for each RT task once. This may be because the pid_task(find_vpid(pid),
PIDTYPE_PID) returns null when the task has not been activated yet.
Since this is the corner case when a task has not been activated, it
should NOT affect the evaluation performance.

Signed-off-by: Meng Xu