linux:module:test VCPUOP_get_rt_info hypercall
A linux module to test the functionality of the added hypercall which
gets the current remaining budget of a specified VCPU.

This is also a showpiece to demonstrate how the hypercall function
should be used.

NB:
    Do NOT directly call the HYPERCALL_ function, use the wrapper
function xen_get_rt_info instead

Test
    Tested and worked
    We can see the period, budget and the current remaining budget of a
specified VCPU in dom0.
    TODO: I have not tested in domU, although it should also work in
domU in my mind experiment

Signed-off-by: Meng Xu