GSN-FPCA2: Solve deadlock issue
We must be very careful if we want to flush cache for a task whene the
task sets its parameters under cache-aware schedulers which will invoke
cache flush for the task as well.
Sicne cache-aware scheduler will flush cache before the task executes,
it is not worth the effort and incur extra overhead if we flush cache
for task when the task set up its rt parameters. So we just do not flush
cache when a task sets its rt paramters under cache aware scheduler

Meng Xu