sys_set_rt_task_param: bug fix for inital cps
Fix the bug when we set the initial cache partitions for a real-time
task.

For cache-aware schedulers, we should not set the initial cache
partitions for rt tasks;
For non-cache-aware schedulers, we can set initial cache partitions for
rt tasks.

I made a mistake when the sys_set_rt_task_param check if the current
scheduler is cache-aware or not.
strcmp(str1, str2) return 0 if the two strings are same.

Signed-off-by: Meng Xu