set_rt_param: flush cache
When we flush cache for the task when the task uses sys_set_rt_param to change its rt paramters, it causes the BUG:

BUG: unable to handle kernel paging request at 0000000000400000

The linear address 400000 has read and execute permission and is mapped to the real-time task binary.

This commit point contain the above bug. We found it on the broadwell machine and will test on haswell machine

Meng Xu