set_rt_param for a task when the task runs
1) We allows a task paramter to be changed when the task is running as a
real-time task.
   This is a hack to allow us to do the experiment for the CAT paper to
show the benefit of dynamic cache management.
   This hack lead to the potential data race on the real-time
parameters. If we change the real-time parameters that are related to
scheduling policy, such as task period, we will affect the schedulng
decision in an *undefined* way.

2) We tested this commit point allows a task to change its own cache
partition configuration when the task is running as a real-time task.
The same code works on the cat-haswell machine but not work on the
cat-broadwell machine.
   I'm not sure why the same code does not work on the cat-broadwell
machine, yet. I will have a look later.

Signed-off-by: Meng Xu