set_rt_task_param: set cos reg
Set cos register when a task (re)sets cache partition configurations, so
that we can see the change of cache partitions promptly.

We cannot rely on the context switch alone to change the cache partition
configuration because the task may keep running which will not trigger
the context switch.

Meng Xu