flush_cache_for_task: shorten critical section
Avoid unnecessarily long critical section when we flush_cache_for_task.
Disable interrupt when we grab the lock.

If we allow preemption when a task is flushing the cache, deadlock may
happen when a high priority preempt a low-priority task who has grabbed
the flush_code_lock and the high-priority task tries to grab the same
lock.

Meng Xu